修复k线无数据切换奔溃

+++ chart/src/main/java/com/jnhyxx/chart/ChartView.java
@@ -520,7 +520,7 @@ public abstract class ChartView extends View {
         return formatNumber(value, mSettings.getNumberScale());
     }
 
-    protected String formatNumber(float value, int numberScale) {
+    protected String formatNumber(double value, int numberScale) {
         DecimalFormat decimalFormat = (DecimalFormat) NumberFormat.getInstance();
 
         decimalFormat.setMaximumFractionDigits(numberScale);
